JVM: Fix non-exist method call (#390)
For unknown reasons, when a random java.lang.Object instance is needed
for method arguments, the LLM model always generates the needed instance
with `FuzzedDataProvider::consumeObject()` or
`FuzzedDataProvider::getObject()`. These two methods are not valid and
never exist. Similar case happens for integer instance. the LLM model
use a non-exist `FuzzedDataProvider::consumeInt(int)` for integer data
generation.

This PR adds a temporary fix to replace the call to these methods with a
valid one.

The `FuzzedDataProvider::consumeObject()` or
`FuzzedDataProvider::getObject()` are replaced by
`FuzzedDataProvider::consumeString(int)` as String is one of the
subclasses for the Object instance.

The `FuzzedDataProvider::consumeInt(int)` is replaced by
`FuzzedDataProvider::consumeInt(int, int)` for getting the correct
internet within range.

---------

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>
Co-authored-by: DavidKorczynski <david@adalogics.com>