Rename gen code to query llm (#480)
The main reason for this is that the `generate_code` does not really
generate any code but rather it queries a given LLM using a specified
prompt. Since we now have prompts of various sort, I feel it might be a
bit misplaced the name. This could also make it a bit more clear which
API to use if you're working on using LLMs for tasks other than explicit
code generation.

This came up while doing https://github.com/google/oss-fuzz-gen/pull/479
where one consideration was to have the LLM generate a corpus explicit
without going a seed-corpus-by-way-of-python generation.

Ref: https://github.com/google/oss-fuzz-gen/issues/482

---------

Signed-off-by: David Korczynski <david@adalogics.com>