JVM: Fix multiple bugs in JVM coverage calculation (#490)
This PR fixes multiple bugs that cause problems in calculating line
coverage diff for the JVM projects.
The bugs are listed below.
1. The path to finding the existing coverage report (jacoco.xml) is
wrong, making the logic always compare with an empty report when
calculating for line coverage diff. This PR adds a new functions to
retrieve and parse the existing coverage report correctly.
2. The interpretation logic for the JVM coverage report on array-type
arguments or empty arguments is wrong, resulting in a non-matching
method name when merging the coverage report. This PR fixes the java
name and argument interpretation to get a correct method signature as
identifiers when combining and subtracting coverage reports.
3. The subtraction of that line diff values between the existing
coverage report and the new coverage report is reversed, making the
calculation wrong. This PR fixes that.
4. The pre-check process of builder_runner.py will check if the target
method exists in the generated harness before building it. The checking
is done by the _contains_target_jvm_method() function. In Java source
code, a single method call could contain an unlimited amount of space
and could also split into multiple lines. The current checking logic
fails to discover if spaces or line feeds exist within a method
invocation. This PR fixes that by pre-processing the code to remove line
feed and space before processing to correctly discover if the target
method is being called or not.
5. Since constructors of classes are some of our targets and there are
too many ways to invoke them, the current logic in
_contains_target_jvm_method() fails to discover existing constructor
invocation and causes false negative cases. This PR skip the pre-check
for the constructor target of the JVM project to avoid those false
negative cases.
6. JVM coverage information is mapped towards random line name but not
the correct line number information. This PR fixes that by modifying the
logic of the from_jvm_file method to correctly interpret the jacoco.xml
report.
7. Subtraction of coverage report only done by comparing the line
coverage count from the coverage reports without checking if different
lines has been covered and result in same covered line count. This PR
fixes the logic of those functions to calculate the coverage gain
correctly.

---------

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>