JVM: Revamp JVM specific properties retrieval for prompt generation (#531)
There are some properties in the benchmarks that only exist for JVM
projects. In the current logic, these properties are being added to the
benchmark class and only JVM projects will use them. This PR reduces the
size of benchmarks by removing these properties from the benchmarks.
Instead, these properties are lively queries from the introspector new
API without the need to store them in the benchmark objects.
In addition to the two properties (exceptions and is_jvm_static) for JVM
projects, a third property is also added to indicate if the target
method is a resource-type object (implements AutoClosable) that needs to
be closed after use. This helps to reduce possible OOM in generated
fuzzing harnesses because of memory leaks from unclosed resource
objects. This is the replacement of the work originally shown in #528.

---------

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>