Fix package missing bug (#606)
The Python script in the `experimental.jvm` package is not intended to
be invoked by the core OFG logic. Instead, it is meant to be executed by
the `run_e2e.sh` script located in the same directory. Therefore, if the
package name `experimental.jvm` needs to be changed as done in #604, the
`sys.path` must be set to the core OFG base path. This PR implements
that change.

---------

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>