Better format agent chat history in reports. (#677)
To make the LLM interactions clearer. Prompts are coloured with a slight
pink background, and responses are moved to the right side with a slight
green background.

Part of #676.

Also stop propagating chat logging to the root logger (to avoid
duplicate output).

This PR also adds some workarounds for when `fixed_targets`,
`raw_targets` don't exist in the results dir for agent runs. This can
happen when we download a results dir from GCS, because empty dirs are
not uploaded.

Example: 


![image](https://github.com/user-attachments/assets/6c04e037-5275-4ff3-90a9-ebbf99098fe9)