[JVM] Add retry logic for no coverage gain (#742)
This PR relocate the calculation of coverage total and coverage diff
into the checking and retry loop. This fixes can then use the lively
caluclated coverage diff to determine if the successfully build and run
harness increase the project coverage or not. If it does not increase
the project coverage, use the new retry logic prompts to ask LLM to help
fixing the harness. The fixing of errors in generated harness as well as
harness with no coverage increase are count together. After this fix, it
increase the rate of successfully build and run for generated harness
and make sure more of them could have help improveing project coverage.

*Remark, this coverage feedback approach is currently only work
exclusively for JVM projects.*

---------

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>