Fix CI failing because of pandas csv fail (#818)
It is observed that the CI starting to failed because pandas.read_csv
will failed wierdly if the file is too large. This PR fixes the read_csv
logic to include the chunksize properties to split the reading into
multiple times and concat them back afterward to avoid the errors.

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>