JVM: Fix wrong use of OpenAI prompt priming and problem  (#899)
This PR fixes the wrong use of OpenAI prompt priming and problem role
for JVM projects. It is understood that, in a chat session of OpenAi,
role `System` is used for setting the base priming, role `user` is used
to store the user prompt and role `assistant` is used to store the
previous reply from OpenAi. Each dictionary in the messages list is act
as the full trace of chat history. This PR fixes the JVM (and
Rust/Python in later PR) approach on the wrong use of these roles in the
OpenAi prompts.

---------

Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>