stage: execution: handle binary data (#1120)
The run logs can contain binary that can't be decoded. This is likely
only a tiny bit of the logs, though, so we shouldn't fail hard on
unicode decoding errors as such. This fixes it by doing a softer
decoding and making sure we don't break because of unicode decode
errors.